[fnd|nstool] Promote IFile wrapped classes to libfnd.
