+++ lib/libfnd/libfnd.vcxproj
@@ -122,6 +122,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="include\fnd\aes.h" />
+    <ClInclude Include="include\fnd\AesCtrWrappedIFile.h" />
     <ClInclude Include="include\fnd\base64.h" />
     <ClInclude Include="include\fnd\BitMath.h" />
     <ClInclude Include="include\fnd\ecdsa.h" />
@@ -131,8 +132,11 @@
     <ClInclude Include="include\fnd\IFile.h" />
     <ClInclude Include="include\fnd\io.h" />
     <ClInclude Include="include\fnd\ISerialisable.h" />
+    <ClInclude Include="include\fnd\LayeredIntegrityMetadata.h" />
+    <ClInclude Include="include\fnd\LayeredIntegrityWrappedIFile.h" />
     <ClInclude Include="include\fnd\List.h" />
     <ClInclude Include="include\fnd\lz4.h" />
+    <ClInclude Include="include\fnd\OffsetAdjustedIFile.h" />
     <ClInclude Include="include\fnd\ResourceFileReader.h" />
     <ClInclude Include="include\fnd\rsa.h" />
     <ClInclude Include="include\fnd\sha.h" />
@@ -144,11 +148,15 @@
     <ClInclude Include="include\fnd\Vec.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="source\AesCtrWrappedIFile.cpp" />
     <ClCompile Include="source\aes_wrapper.cpp" />
     <ClCompile Include="source\base64_wrapper.cpp" />
     <ClCompile Include="source\Exception.cpp" />
     <ClCompile Include="source\io.cpp" />
+    <ClCompile Include="source\LayeredIntegrityMetadata.cpp" />
+    <ClCompile Include="source\LayeredIntegrityWrappedIFile.cpp" />
     <ClCompile Include="source\lz4_wrapper.cpp" />
+    <ClCompile Include="source\OffsetAdjustedIFile.cpp" />
     <ClCompile Include="source\ResourceFileReader.cpp" />
     <ClCompile Include="source\rsa_wrapper.cpp" />
     <ClCompile Include="source\sha_wrapper.cpp" />
