+++ programs/nstool/source/CnmtProcess.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <iomanip>
 #include <fnd/SimpleTextOutput.h>
-#include "OffsetAdjustedIFile.h"
+#include <fnd/OffsetAdjustedIFile.h>
 #include "CnmtProcess.h"
 
 CnmtProcess::CnmtProcess() :
