+++ programs/nstool/source/NacpProcess.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <iomanip>
 #include <fnd/SimpleTextOutput.h>
-#include "OffsetAdjustedIFile.h"
+#include <fnd/OffsetAdjustedIFile.h>
 #include "NacpProcess.h"
 
 NacpProcess::NacpProcess() :
