+++ programs/nstool/source/NsoProcess.cpp
@@ -1,9 +1,9 @@
 #include <iostream>
 #include <iomanip>
 #include <fnd/SimpleTextOutput.h>
+#include <fnd/OffsetAdjustedIFile.h>
 #include <fnd/Vec.h>
 #include <fnd/lz4.h>
-#include "OffsetAdjustedIFile.h"
 #include "NsoProcess.h"
 
 NsoProcess::NsoProcess():
