+++ programs/nstool/source/PkiCertProcess.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 #include <iomanip>
 #include <fnd/SimpleTextOutput.h>
+#include <fnd/OffsetAdjustedIFile.h>
 #include <nn/pki/SignUtils.h>
-#include "OffsetAdjustedIFile.h"
 #include "PkiCertProcess.h"
 #include "PkiValidator.h"
